Working on _memoize
I'm having trouble getting this one down. I understand that I've got to
'cache' the results of a function and return the cached result if the
same function (from the same execution context). It seems to be doing
that, however, if I'm calling the function from a new execution context,
it's not returning the proper result.

-JEE